# swagger.yaml
openapi: 3.0.0
info:
  title: Node.js Boilerplate API
  description: Authentication API with JWT, Refresh Tokens and RBAC
  version: 1.0.0

servers:
  - url: http://localhost:4000
    description: Local development server

tags:
  - name: Accounts
    description: Account management and authentication

paths:
  /accounts/authenticate:
    post:
      summary: Authenticate account credentials and return a JWT token
      description: Accounts must be verified before authenticating.
      operationId: authenticate
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  example: admin@example.com
                password:
                  type: string
                  example: "pass123"
      responses:
        '200':
          description: Account details, a JWT access token and a refresh token cookie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

  /accounts/refresh-token:
    post:
      summary: Use a refresh token to get a new JWT token
      operationId: refreshToken
      tags: [Accounts]
      responses:
        '200':
          description: Account details, a new JWT access token and new refresh token cookie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

  /accounts/revoke-token:
    post:
      summary: Revoke a refresh token
      operationId: revokeToken
      tags: [Accounts]
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        '200':
          description: Token revoked
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /accounts/register:
    post:
      summary: Register a new user account
      operationId: register
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registration successful
        '400':
          $ref: '#/components/responses/BadRequest'

  /accounts/verify-email:
    post:
      summary: Verify a new account with a verification token
      operationId: verifyEmail
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        '200':
          description: Verification successful
        '400':
          $ref: '#/components/responses/BadRequest'

  /accounts/forgot-password:
    post:
      summary: Submit email address to reset the password on an account
      operationId: forgotPassword
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Check email for reset instructions

  /accounts/validate-reset-token:
    post:
      summary: Validate the reset password token received by email
      operationId: validateResetToken
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        '200':
          description: Token is valid
        '400':
          $ref: '#/components/responses/BadRequest'

  /accounts/reset-password:
    post:
      summary: Reset the password for an account
      operationId: resetPassword
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset successful
        '400':
          $ref: '#/components/responses/BadRequest'

  /accounts:
    get:
      summary: Get a list of all accounts (Admin only)
      operationId: getAll
      tags: [Accounts]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Array of all accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Create a new account (Admin only)
      operationId: createAccount
      tags: [Accounts]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
      responses:
        '200':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /accounts/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
    get:
      summary: Get a single account by id
      operationId: getAccountById
      tags: [Accounts]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Account details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update an account
      operationId: updateAccount
      tags: [Accounts]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAccountRequest'
      responses:
        '200':
          description: Account updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      summary: Delete an account
      operationId: deleteAccount
      tags: [Accounts]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Account deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    AccountResponse:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        role:
          type: string
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
        isVerified:
          type: boolean

    AuthenticateResponse:
      allOf:
        - $ref: '#/components/schemas/AccountResponse'
        - type: object
          properties:
            jwtToken:
              type: string

    RegisterRequest:
      type: object
      required: [firstName, lastName, email, password, confirmPassword, acceptTerms]
      properties:
        title:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        confirmPassword:
          type: string
        acceptTerms:
          type: boolean

    CreateAccountRequest:
      type: object
      required: [firstName, lastName, email, password, confirmPassword, role]
      properties:
        title:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        confirmPassword:
          type: string
        role:
          type: string
          enum: [Admin, User]

    UpdateAccountRequest:
      type: object
      properties:
        title:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        confirmPassword:
          type: string
        role:
          type: string
          enum: [Admin, User]

    ResetPasswordRequest:
      type: object
      required: [token, password, confirmPassword]
      properties:
        token:
          type: string
        password:
          type: string
          minLength: 6
        confirmPassword:
          type: string

  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string